// src/lib/models/add-contact-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API Documentation
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface AddContactDto
 */
export interface AddContactDto {

    /**
     * @type {string}
     * @memberof AddContactDto
     */
    requesterId: string;

    /**
     * @type {string}
     * @memberof AddContactDto
     */
    contactId: string;

    /**
     * @type {string}
     * @memberof AddContactDto
     */
    projectId: string;
}